+++ e_contact/management/commands/e_contact_notify.py
@@ -1,0 +1,9 @@
+import djclick as click
+
+from e_contact.utils import notify
+
+
+@click.command()
+def command():
+    notified = notify()
+    print(f'Notified? {notified}')